updated packages, plugins and scripts

+++ wp-content/plugins/wordpress-seo/admin/ajax.php
@@ -45,6 +45,11 @@ function wpseo_set_option() {
 add_action( 'wp_ajax_wpseo_set_option', 'wpseo_set_option' );
 
 /**
+ * Since 3.2 Notifications are dismissed in the Notification Center.
+ */
+add_action( 'wp_ajax_yoast_dismiss_notification', array( 'Yoast_Notification_Center', 'ajax_dismiss_notification' ) );
+
+/**
  * Function used to remove the admin notices for several purposes, dies on exit.
  */
 function wpseo_set_ignore() {
@@ -64,23 +69,6 @@ function wpseo_set_ignore() {
 }
 
 add_action( 'wp_ajax_wpseo_set_ignore', 'wpseo_set_ignore' );
-
-/**
- * Hides the after-update notification until the next update for a specific user.
- */
-function wpseo_dismiss_about() {
-	if ( ! current_user_can( 'manage_options' ) ) {
-		die( '-1' );
-	}
-
-	check_ajax_referer( 'wpseo-dismiss-about' );
-
-	update_user_meta( get_current_user_id(), 'wpseo_seen_about_version' , WPSEO_VERSION );
-
-	die( '1' );
-}
-
-add_action( 'wp_ajax_wpseo_dismiss_about', 'wpseo_dismiss_about' );
 
 /**
  * Hides the default tagline notice for a specific user.
@@ -109,23 +97,35 @@ function wpseo_kill_blocking_files() {
 
 	check_ajax_referer( 'wpseo-blocking-files' );
 
-	$message = 'There were no files to delete.';
+	$message = 'success';
+	$errors  = array();
+
+	// Todo: Use WP_Filesystem, but not so easy to use in AJAX with credentials form still internal.
 	$options = get_option( 'wpseo' );
 	if ( is_array( $options['blocking_files'] ) && $options['blocking_files'] !== array() ) {
-		$message       = 'success';
-		$files_removed = 0;
-		foreach ( $options['blocking_files'] as $k => $file ) {
-			if ( ! @unlink( $file ) ) {
-				$message = __( 'Some files could not be removed. Please remove them via FTP.', 'wordpress-seo' );
+		foreach ( $options['blocking_files'] as $file ) {
+			if ( is_file( $file ) ) {
+				if ( ! @unlink( $file ) ) {
+					$errors[] = __(
+						sprintf( 'The file "%s" could not be removed. Please remove it via FTP.', $file ),
+						'wordpress-seo'
+					);
+				}
 			}
-			else {
-				unset( $options['blocking_files'][ $k ] );
-				$files_removed ++;
+
+			if ( is_dir( $file ) ) {
+				if ( ! @ rmdir( $file ) ) {
+					$errors[] = __(
+						sprintf( 'The directory "%s" could not be removed. Please remove it via FTP.', $file ),
+						'wordpress-seo'
+					);
+				}
 			}
 		}
-		if ( $files_removed > 0 ) {
-			update_option( 'wpseo', $options );
-		}
+	}
+
+	if ( $errors ) {
+		$message = implode( '<br />', $errors );
 	}
 
 	die( $message );
@@ -322,6 +322,9 @@ function wpseo_upsert_new( $what, $post_id, $new, $original ) {
  * Create an export and return the URL
  */
 function wpseo_get_export() {
+	if ( ! current_user_can( 'manage_options' ) ) {
+		die( '-1' );
+	}
 
 	$include_taxonomy = ( filter_input( INPUT_POST, 'include_taxonomy' ) === 'true' );
 	$export           = new WPSEO_Export( $include_taxonomy );
@@ -355,6 +358,10 @@ function ajax_get_keyword_usage() {
 	$post_id = filter_input( INPUT_POST, 'post_id' );
 	$keyword = filter_input( INPUT_POST, 'keyword' );
 
+	if ( ! current_user_can( 'edit_post', $post_id ) ) {
+		die( '-1' );
+	}
+
 	wp_die(
 		WPSEO_Utils::json_encode( WPSEO_Meta::keyword_usage( $keyword, $post_id ) )
 	);
@@ -370,12 +377,54 @@ function ajax_get_term_keyword_usage() {
 	$keyword = filter_input( INPUT_POST, 'keyword' );
 	$taxonomy = filter_input( INPUT_POST, 'taxonomy' );
 
+	if ( ! current_user_can( 'edit_terms' ) ) {
+		die( '-1' );
+	}
+
+	$usage = WPSEO_Taxonomy_Meta::get_keyword_usage( $keyword, $post_id, $taxonomy );
+
+	// Normalize the result so it it the same as the post keyword usage AJAX request.
+	$usage = $usage[ $keyword ];
+
 	wp_die(
-		WPSEO_Utils::json_encode( WPSEO_Taxonomy_Meta::get_keyword_usage( $keyword, $post_id, $taxonomy ) )
+		WPSEO_Utils::json_encode( $usage )
 	);
 }
 
 add_action( 'wp_ajax_get_term_keyword_usage',  'ajax_get_term_keyword_usage' );
+
+/**
+ * Removes stopword from the sample permalink that is generated in an AJAX request
+ *
+ * @param array  $permalink The permalink generated for this post by WordPress.
+ * @param int    $post_ID The ID of the post.
+ * @param string $title The title for the post that the user used.
+ * @param string $name The name for the post that the user used.
+ *
+ * @return array
+ */
+function wpseo_remove_stopwords_sample_permalink( $permalink, $post_ID, $title, $name ) {
+	WPSEO_Options::get_instance();
+	$options = WPSEO_Options::get_options( array( 'wpseo_permalinks' ) );
+	if ( $options['cleanslugs'] !== true ) {
+		return $permalink;
+	}
+
+	/*
+	 * If the name is empty and the title is not, WordPress will generate a slug. In that case we want to remove stop
+	 * words from the slug.
+	 */
+	if ( empty( $name ) && ! empty( $title ) ) {
+		$stop_words = new WPSEO_Admin_Stop_Words();
+
+		// The second element is the slug.
+		$permalink[1] = $stop_words->remove_in( $permalink[1] );
+	}
+
+	return $permalink;
+}
+
+add_action( 'get_sample_permalink', 'wpseo_remove_stopwords_sample_permalink', 10, 4 );
 
 // Crawl Issue Manager AJAX hooks.
 new WPSEO_GSC_Ajax;
@@ -388,6 +437,9 @@ new Yoast_Dashboard_Widget();
 new Yoast_OnPage_Ajax();
 
 new WPSEO_Shortcode_Filter();
+
+new WPSEO_Taxonomy_Columns();
+
 
 // Setting the notice for the recalculate the posts.
 new Yoast_Dismissable_Notice_Ajax( 'recalculate', Yoast_Dismissable_Notice_Ajax::FOR_SITE );
